32-bit float support
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@5896 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DICOMExport.m
@@ -126,8 +126,8 @@
 }
 
 - (long) setPixelData:		(unsigned char*) idata
-		samplesPerPixel:	(long) ispp
-		bitsPerSample:		(long) ibps
+		samplesPerPixel:	(int) ispp
+		bitsPerSample:		(int) ibps
 		width:				(long) iwidth
 		height:				(long) iheight
 {
